Support subframe navigation to urn:uuid URLs in Subresource Web Bundles
Design doc: https://bit.ly/3kLkMRJ#heading=h.aay60dv2vzua

This CL changes BeginNavigation IPC to take an additional parameter,
URNLoaderParams. Currently its only member is a
URLLoaderClientEndpoints, which is used to load a urn:uuid resource from
a Web Bundle.

In the renderer side, FrameLoader::StartNavigation calls
SubresourceWebBundle::MaybeCreateWebBundleBackedNavigationInfo() and it
creates a URLLoaderClientEndpoints if the request URL is a urn:uuid URL
in the bundle.

In the browser process side, if URNLoaderParams is available
NavigationURLLoaderImpl creates a UrnURLLoaderFactory with the
endpoints, and sets it as the handler for urn: scheme.

At the moment, browser process doesn't care which bundle the urn:uuid
resource came from. In followup CLs, we will add information about
WebBundle URL to URNLoaderParams, so that cache isolation can be done
taking the WebBundle's origin into account.

This CL makes CanAccessDataForOrigin allow urn: lock URLs when using
CanAccessDataForOrigin to validate precursor of an opaque origin, as we
do for data: URLs.

Bug: 1082020
Change-Id: Ia209f71bcd6855df29581d87595010acd609e8ee